+++ etherparse/src/internet/mod.rs
@@ -1,10 +1,10 @@
-pub mod internet_slice;
 pub mod ip_auth_header;
 pub mod ip_auth_header_slice;
 pub mod ip_frag_offset;
 pub mod ip_header;
 pub mod ip_number_impl;
 pub mod ip_payload;
+pub mod ip_slice;
 pub mod ipv4_dscp;
 pub mod ipv4_ecn;
 pub mod ipv4_exts;